ODE-1264 try again tom make it work in Travis
